Переносит код блока комментариев из отдельного компонента в модуль класса FilmDetails.
